refactor: create ad error modal
