+++ packages/p2p/src/components/modal-manager/modals/create-ad-error-modal/index.ts
@@ -1,4 +1,4 @@
-import CreateAdErrorModal from './create-ad-error-modal.jsx';
+import CreateAdErrorModal from './create-ad-error-modal';
 import './create-ad-error-modal.scss';
 
 export default CreateAdErrorModal;
